refactor: 폴더 구조 나누기
