+++ android/app/src/androidTest/java/com/pbl/grandmarket_android/LoginActivityTest.kt
@@ -7,6 +7,7 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.pbl.grandmarket_android.ui.login.LoginActivity
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
