--- android/app/src/main/java/com/pbl/grandmarket_android/BaseActivity.kt
@@ -1,24 +1,0 @@
-package com.pbl.grandmarket_android
-
-import android.os.Bundle
-import android.view.View
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowCompat
-import androidx.core.view.WindowInsetsCompat
-
-abstract class BaseActivity: AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        WindowCompat.setDecorFitsSystemWindows(window,false)    //edge-to-edge 활성화
-    }
-
-    protected fun applyBottomInsets(view: View) {
-        val initialPaddingBottom = view.paddingBottom
-        ViewCompat.setOnApplyWindowInsetsListener(view) { v, insets ->
-            val navBars = insets.getInsets(WindowInsetsCompat.Type.navigationBars())
-            v.setPadding(v.paddingLeft, v.paddingTop, v.paddingRight, navBars.bottom+initialPaddingBottom)
-            insets
-        }
-    }
-}
