--- android/app/src/main/java/com/pbl/grandmarket_android/UserRole.kt
@@ -1,12 +1,0 @@
-package com.pbl.grandmarket_android
-
-enum class UserRole(val value: String) {
-    SELLER("SELLER"),
-    BUYER("BUYER");
-
-    companion object {
-        fun from(value: String?): UserRole {
-            return values().firstOrNull { it.value == value } ?: BUYER
-        }
-    }
-}
