+++ android/app/src/main/res/layout/fragment_bottom_sheet_store_list.xml
