+++ android/app/src/main/res/layout/fragment_map_seller.xml
